TSSEscore {ATACseqQC}	R Documentation

Transcription Start Site (TSS) Enrichment Score

Description

TSS score is a raio between aggregate distribution of reads centered on TSSs and that flanking the corresponding TSSs. TSS score = the depth of TSS (each step within 1000 bp each side) / the depth of end flanks (100bp each end). TSSE score = max(mean(TSS score in each window)).

Usage


TSSEscore(
  obj,
  txs,
  seqlev = intersect(seqlevels(obj), seqlevels(txs)),
  upstream = 1000,
  downstream = 1000,
  endSize = 100,
  width = 100,
  step = width,
  pseudocount = 0
)

Arguments

obj
an object of GAlignments

txs
GRanges of transcripts

seqlev
A vector of characters indicates the sequence levels.

upstream, downstream
numeric(1) or integer(1). upstream and downstream of TSS. Default is 1000

endSize
numeric(1) or integer(1). the size of the end flanks. Default is 100

width
numeric(1) or integer(1). the window size for TSS score. Default is 100.

step
numeric(1) or integer(1). The distance between the start position of the sliding windows.

pseudocount
numeric(1) or integer(1). Pseudocount. Default is 0. If pseudocount is no greater than 0, the features with ZERO or less than ZERO counts in flank region will be removed in calculation.

Value

A object of GRanges with TSS scores

Author(s)

Jianhong Ou

References

https://www.encodeproject.org/data-standards/terms/#enrichment

Examples


 
library(GenomicRanges)
bamfile <- system.file("extdata", "GL1.bam", 
                       package="ATACseqQC", mustWork=TRUE)
gal1 <- readBamFile(bamFile=bamfile, tag=character(0), 
                    which=GRanges("chr1", IRanges(1, 1e6)), 
                    asMates=FALSE)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
txs <- transcripts(TxDb.Hsapiens.UCSC.hg19.knownGene)
tsse <- TSSEscore(gal1, txs)

[Package ATACseqQC version 1.16.0 Index]