distanceDyad {ATACseqQC}	R Documentation

Distance of potential nucleosome dyad

Description

Calculate the distance of potential nucleosome dyad and the linear model for V.

Usage


distanceDyad(vPlotOut, fragLenRanges = c(60, 180, 250), draw = TRUE, ...)

Arguments

vPlotOut
The output of vPlot.

fragLenRanges
A numeric vector (length=3) for fragment size of nucleosome free and mono-nucleosome. Default c(60, 180, 250).

draw
Plot the results or not. Default TRUE.

...
Prameters could be passed to plot.

Value

an invisible list with distance of nucleosome and the linear model.

Author(s)

Jianhong Ou

See Also

vPlot

Examples


bamfile <- system.file("extdata", "GL1.bam",
                       package="ATACseqQC")
library(MotifDb)
CTCF <- query(MotifDb, c("CTCF"))
CTCF <- as.list(CTCF)
library(BSgenome.Hsapiens.UCSC.hg19)
vp <- vPlot(bamfile, pfm=CTCF[[1]],
     genome=Hsapiens,
     min.score="95%", seqlev="chr1",
     draw=FALSE)
distanceDyad(vp)
     

[Package ATACseqQC version 1.16.0 Index]