fragSizeDist {ATACseqQC}	R Documentation

fragment size distribution

Description

estimate the fragment size of bams

Usage


fragSizeDist(
  bamFiles,
  bamFiles.labels,
  index = bamFiles,
  ylim = NULL,
  logYlim = NULL
)

Arguments

bamFiles
A vector of characters indicates the file names of bams.

bamFiles.labels
labels of the bam files, used for pdf file naming.

index
The names of the index file of the 'BAM' file being processed; This is given without the '.bai' extension.

ylim
numeric(2). ylim of the histogram.

logYlim
numeric(2). ylim of log-transformed histogram for the insert.

Value

Invisible fragment length distribution list.

Author(s)

Jianhong Ou

Examples


bamFiles <- dir(system.file("extdata", package="ATACseqQC"), "GL.*.bam$", full.names=TRUE)
bamFiles.labels <- sub(".bam", "", basename(bamFiles))
fragSizeDist(bamFiles, bamFiles.labels)

[Package ATACseqQC version 1.16.0 Index]