pwmscores {ATACseqQC}	R Documentation

max PWM scores for sequences

Description

calculate the maximal PWM scores for each given sequences

Usage


pwmscores(pwm, subject)

Arguments

pwm
A Position Weight Matrix represented as a numeric matrix with row names A, C, G and T.

subject
Typically a DNAString object. A Views object on a DNAString subject, a MaskedDNAString object, or a single character string, are also supported. IUPAC ambiguity letters in subject are ignored (i.e. assigned weight 0) with a warning.

Value

a numeric vector

Author(s)

Jianhong

[Package ATACseqQC version 1.16.0 Index]